Final Commit

+++ app/src/main/java/com/funfit/usjr/thesis/funfitv2/model/RequestRun.java
@@ -4,15 +4,15 @@ import java.io.Serializable;
 
 public class RequestRun implements Serializable{
 	
-	private int runId;
+	private long runId;
 	private String date;
-	private int distance;
-	private int time;
-	private int userId;
+	private long distance;
+	private long time;
+	private long userId;
 	
 	public RequestRun(){}
 
-	public RequestRun(int runId, String date, int distance, int time, int userId) {
+	public RequestRun(long runId, String date, long distance, long time, long userId) {
 		super();
 		this.runId = runId;
 		this.date = date;
@@ -21,11 +21,11 @@ public class RequestRun implements Serializable{
 		this.userId = userId;
 	}
 
-	public int getRunId() {
+	public long getRunId() {
 		return runId;
 	}
 
-	public void setRunId(int runId) {
+	public void setRunId(long runId) {
 		this.runId = runId;
 	}
 
@@ -37,27 +37,27 @@ public class RequestRun implements Serializable{
 		this.date = date;
 	}
 
-	public int getDistance() {
+	public long getDistance() {
 		return distance;
 	}
 
-	public void setDistance(int distance) {
+	public void setDistance(long distance) {
 		this.distance = distance;
 	}
 
-	public int getTime() {
+	public long getTime() {
 		return time;
 	}
 
-	public void setTime(int time) {
+	public void setTime(long time) {
 		this.time = time;
 	}
 
-	public int getUserId() {
+	public long getUserId() {
 		return userId;
 	}
 
-	public void setUserId(int userId) {
+	public void setUserId(long userId) {
 		this.userId = userId;
 	}
 }
